add-dog - move general-dog-form into my-pets/_components
