+++ src/app/user/(logged-in)/my-pets/add-dog/_components/add-dog-form.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { BarkFormOption } from "@/components/bark/bark-form";
-import GeneralDogForm, { DogFormData } from "./general-dog-form";
+import GeneralDogForm, { DogFormData } from "../../_components/general-dog-form";
 
 export default function AddDogForm(props: { vetOptions: BarkFormOption[] }) {
   const { vetOptions } = props;
